Restore implicit scrolling behavior
When switching to scrollIntoView (standard) from
scrollIntoViewIfNeeded (non-standard), we lost conditionally scrolling. Add to
the fact that scrollIntoView will default to top of page can cause some hidden
element issues (since some pages like to have sticky headers).

This changes the defaults of scrollIntoView to center by default and opt not to
scroll if the element is already within the window viewport bounds.

Co-authored-by: Justin Watt <justinw@mayvenn.com>